--- test/types/index.test-d.ts
@@ -1,16 +1,0 @@
-import { expectType } from "tsd";
-import SerializerSelector, {
-  SerializerCompiler,
-  SerializerSelector as SerializerSelectorNamed,
-  StandaloneSerializer,
-} from "../..";
-
-{
-  const compiler = SerializerSelector();
-  expectType<SerializerCompiler>(compiler);
-}
-
-{
-  const compiler = SerializerSelectorNamed();
-  expectType<SerializerCompiler>(compiler);
-}
